Moved video cache info widget to it's own file which cleans up the videoCache class a bit. Also added ability to hide the detailed caching info.
