+++ ui/mainwindow.ui
@@ -243,7 +243,7 @@
    <attribute name="dockWidgetArea">
     <number>1</number>
    </attribute>
-   <widget class="QWidget" name="dockWidgetContents"/>
+   <widget class="VideoCacheInfoWidget" name="cachingInfoWidget"/>
   </widget>
   <widget class="QDockWidget" name="displayDockWidget">
    <property name="sizePolicy">
@@ -310,6 +310,12 @@
    <class>PropertiesWidget</class>
    <extends>QWidget</extends>
    <header>propertiesWidget.h</header>
+   <container>1</container>
+  </customwidget>
+  <customwidget>
+   <class>VideoCacheInfoWidget</class>
+   <extends>QWidget</extends>
+   <header>videoCacheInfoWidget.h</header>
    <container>1</container>
   </customwidget>
  </customwidgets>
